# api/openapi.yaml
openapi: 3.0.3
info:
  title: Schedularr API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /blocks/import:
    post:
      operationId: importBlocks
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      parameters:
        - name: dry_run
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
...
  /cron/next:
    get:
      operationId: nextCronOccurrences
      summary: Upcoming occurrences of a cron expression
      description: >-
        The next occurrences of `expr`, evaluated server-side in the
        configured `log.timezone`. It exists so the client never
        re-implements calendar semantics: DST transitions, month lengths,
        and descriptor handling all belong to one parser, the same one the
        scheduling engine uses.

        Occurrences are strictly AFTER `from`, so a `from` sitting exactly
        on an occurrence does not return it again. The array may be shorter
        than `count`: a well-formed expression that never fires (February
        30th) returns an empty array rather than an error, because the
        expression is valid and the answer is genuinely "never".

        Stateless -- it reads no blocks and touches no store.
      parameters:
        - name: expr
          in: query
          required: true
          schema:
            type: string
          description: A standard 5-field cron expression, or a descriptor such as `@daily`.
        - name: count
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 instant to search after. Defaults to now.
      responses:
        "200":
          description: The upcoming occurrences, oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [occurrences]
                properties:
                  occurrences:
                    type: array
                    items:
                      type: string
                      format: date-time
        "400":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
...
    ImportResult:
      type: object
      required: [imported, dry_run]
      properties:
        imported:
          type: integer
        dry_run:
          type: boolean
        names:
          type: array
          items:
            type: string
...